+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CRUDlex Changelog
 Released: Upcoming
 - Added the possibility to override the exact templates of specific pages and not just only the layouts
 - Initializing all needed providers in the CRUDServiceProvider if not done yet by the application
+- Made the nameField of a reference optional
 
 ## 0.9.7
 Released: 2015-07-26